requirements.txt
pydantic~=2.10
requests
psycopg2-binary
schedule
geopy